Change `let ?x = ...` to use the `IP` class.
XXX: Technically, the IP bindings that define the values of the implicit
parameters should be cast to a dictionary for the class (see `mkIPBox` in
module `MkCore`).  At the moment we don't do this, which is not quite
right (although things still work because the value and the dictionary
are represented in the same way).