Restore custom handling of IP for ambiguity (due to a fun-dep bug in GHC).
GHC considers types like these to be ambiguous:

f :: IP "c" a => String

This is not quite right, and the explanation of why that happens
(from module FunDeps) also seems to be not quite correct: all variables
appearing in a type and not showing up in the environment should
be generalized, even though they might be fully determined from nothing
(as above, for example).

Until we fix that problem, however, we keep the special rules for
ambiguity and implicit parameters.