Checkpointing some makefile simplifications that I attempted.
This is not in a working state yet, as we are trying to build
something that goes into the build/ dir but it is not yet
created. Not sure how to add dependency tracking properly.

Btw., Ian is working in this area on the master branch so
there will be a bunch of conflicts anyway :-(