Rearrange the typechecking of arrows, especially arrow "forms"
The typechecking of arrow forms (in GHC 7.6) is known to be bogus,
because it marches down tuple types that may not yet be fully worked
out, depending on when constraint solving happens.  Moreover,
coercions are generated and simply discarded.  The fact that it works
at all is a miracle.

This refactoring is based on a conversation with Ross, where we
rearranged the typing of the argument stack, so that the arrows
have the form
   a (env, (arg1, (arg2, ...(argn, ())))) res
rather than
   a (arg1, (arg2, ...(argn, env))) res
as it was before.

This is vastly simpler to typecheck; just look at the beautiful,
simple type checking of arrow forms now!

We need a new HsCmdCast to capture the coercions generated from
the argument stack.

Current state of play:
   * syntax done
   * renamer done
   * typechecker done
   * desugarer NOT DONE

So nothing will work yet, hence on a branch