Checkpoint: generalize to support arbitrary built-in constructors.
This is done by adding a new TyConRHS (cf TyCon).
The operations for type-checking a built-in tycon are in FamInst.

This is a checkpoint because the current build does not seem to work
on my machine: for some reason things like (+) are considered to be Open.

The reason could be that I need to rebuild more things or, perhaps the
type nat tycons need to be introduced in a different way.  Currently,
they are just declared as ordinary type families.