Implement -XStaticValues V2.
Summary:
Contains contributions from Alexander Vershilov and Mathieu Boespflug.

As proposed in [1], this extension introduces a new syntactic form
`static e`, where `e :: a` can be any closed expression. The static form
produces a value of type `StaticPtr a`, which works as a reference that
programs can "dereference" to get the value of `e` back. References are
like `Ptr`s, except that they are stable across invocations of a
program.

In essence the extension collects the arguments of the static form into
a global static pointer table. The expressions can be looked up by a
fingerprint computed from the package, the module and a fresh name
given to the expression. For more details we refer to the users guide
section contained in the patch.

The extension is a contribution to the Cloud Haskell ecosystem
(distributed-process and related), and thus has the potential to foster
Haskell as a programming language for distributed systems.

The immediate improvement brought by the extension is the elimination of
remote tables from Cloud Haskell applications. Such applications contain
table fragments spread throughout multiple modules and packages.
Eliminating these fragments saves the programmer the burden required to
construct and assemble the global remote table, a verbose and
error-prone process, even with the help of Template Haskell, that
moreover pollutes the export lists of all modules.

The relevant wiki pages are [2, 3]. With respect to the wiki pages, we are not
implementing floating of expressions yet. So `static e` only allows the body to
be an identifier.

This is work in progress and we would be thankful for feedback on how to solve
some technical issues. A presentation of the issues follows.

[1] Jeff Epstein, Andrew P. Black, and Simon Peyton-Jones. Towards
Haskell in the cloud. SIGPLAN Not., 46(12):118–129, September 2011. ISSN
0362-1340.
[2] https://ghc.haskell.org/trac/ghc/wiki/StaticPointers
[3] https://ghc.haskell.org/trac/ghc/wiki/StaticPointers/ImplementationPlan

The typing rule for static
--------------------------

This rule has been suggested for static:

    Top(Gamma); empty |- e :: tau
    -----------------------------------
    Gamma; Theta |- static e :: StaticPtr taut

where `Gamma` gives the types of all in-scope variables (e.g. `a:*`, `x::a->a`),
`Theta` gives the in-scope given constraints (e.g. `Eq a`) and `tau` is the type
of `e`.

When typechecking `static e` we typecheck `e` in the *top-level environment
`Top(Gamma)`, i.e. discarding all locally bound variables (eg lambda-bound).
And we want to discard `Theta`, so that the only constraints we can use in
solving `e` are the top-level instances.

With this rule, the following program would fail to typecheck:

    {-# LANGUAGE StaticPointers #-}
    {-# LANGUAGE GADTs #-}

    import GHC.StaticPtr
    import Data.Typeable

    main :: IO ()
    main = print $ h 'a'

    h :: (Typeable b, Char ~ b) => b -> String
    h a = deRefStaticPtr (static show) a

Typechecking for show cannot use the constraint `Char ~ b`, therefore its type
is inferred as `Show a => a -> String`, which is not `Typeable`.

A rule that maybe would allow such program to typecheck is:

      Top(Gamma); Theta |- e :: tau
    ----------------------------------------- wanted(e) and Theta are disjoint
    Gamma; Theta |- static e :: StaticPtr tau

`e` doesn't use constraints in Theta, but nonetheless the constraints in scope
could affect how variables are unified. Thus show can be given the type
`Char -> String`.

Is the later rule acceptable? This is the rule the current patch implements.

`checkStaticPointer` in TcRnDriver.hs
-------------------------------------

Probably, the function `checkStaticPointer`, defined in TcRnFriver.hs should be
moved or reimplemented under `simplifyTop`, since it is solving constraints for
the bodies of static forms. However we don’t have a clear idea on which
functions should be modified or which functions should be used.

A coercion around the static body appears if omitting a type signature
----------------------------------------------------------------------

    $ cat t.hs
    {-# LANGUAGE StaticPointers #-}
    import GHC.StaticPtr

    foo :: Int
    foo = 0
    main = print $ deRefStaticPtr (static foo)

The above program produces the following output:

    $ /home/facundo/tweag/ghc/inplace/bin/ghc-stage2 --make t.hs -fforce-recomp -dcore-lint -dppr-debug
    [1 of 1] Compiling Main             ( t2.hs, t2.o )

    t2.hs:8:32:
       The argument of a static form can be only a name
       but found: static ({t2.hs:8:39-41}
                      main@main:Main.foo{v rmm} |> (Sub cobox_aAv{v}[lid]))

But it compiles fine if a type signature is added:

    main = print $ deRefStaticPtr (static foo :: StaticPtr Int)

We don’t know yet how to prescind of the type signature.

Assertion related to untouchables appears if omitting a type signature
----------------------------------------------------------------------

    $ cat t.hs
    {-# LANGUAGE DeriveDataTypeable #-}
    {-# LANGUAGE StaticPointers #-}
    import GHC.StaticPtr
    import Data.Typeable

    data T a = T { t_field :: a }
      deriving Typeable
    main = print $ deRefStaticPtr (static t_field) $ T 0

The above program produces the following output:

    $ /home/facundo/tweag/ghc/inplace/bin/ghc-stage2 --make t2.hs -fforce-recomp -ddump-foreign -dcore-lint
    [1 of 1] Compiling Main             ( t2.hs, t2.o )
    ghc-stage2: panic! (the 'impossible' happened)
      (GHC version 7.9.20141128 for x86_64-unknown-linux):
    	ASSERT failed!
      file compiler/typecheck/TcType.lhs line 645
      a_aV2[tau:3]
      3
      2

    Please report this as a GHC bug:  http://www.haskell.org/ghc/reportabug

But it builds fine if a type signature is added:

    main = print $ deRefStaticPtr (static t_field :: StaticPtr (T Char -> Char)) $ T ‘a’

We don’t know yet how to prescind of the type signature.

Test Plan: validate

Reviewers: austin, simonpj, simonmar, hvr

Subscribers: carter, thomie, goldfire

Differential Revision: https://phabricator.haskell.org/D550

GHC Trac Issues: #7015