Laser scaling
See #43

We now aim out lasers precisely and realise that adding a scaling to MG
is the wrong solution.

think mg_weight is a big hammer hitting the wrong nail. Looking closely at the code the correct information about the weight comes from mg_arg_tys which marks the weight of each pattern appropriately.

This all nearly goes so right but fails with a flourish as in selectMatchVar, the weight and variable come very close to each other before the weight is ignored and then then all variables scaled inappropriately by the mg_weight hammer in an outer scope.

After all, if we inspect how the multiplicities of mg_arg_tys come about, they are calculated from the type of function definition. Arnaud already slotted through this information but I failed to complete the dunk on the other end.

Fix: Scale at the call to localiseId and write a note explaining this as it is subtle.