Remove aliases
In Lint, every nonrecursive let adds the name to le_ue_aliases.
We consider a variable to be an alias iff it's present in le_ue_aliases.

When binding a variable in a case or lambda, we need to remove it
from le_ue_aliases. This caused an issue in code with shadowing:

let x = True in case ... of x[1] -> x

When checking the last 'x', the outer x was present in le_ue_aliases
and this caused the occurrence of the inner x not to be recorded.
Solution: inside every binder, we call deleteAlias to remove the
shadowed variable.

Refs #357