Fix multiple scaling issues
This fixes #212, data T p a where MkT :: a -->.(p) T p a.

There were two bugs.
1. During kind-checking of a datacon, we didn't
call tcMult on field multiplicities. As a result,
T was presumed to be * -> * -> * which caused a problem
during type-checking.

2. The pattern of using Compose with Traversable instance
for Scaled is completely wrong.

A call such as

mapM zonk (Compose [Scaled m1 t1, Scaled m2 t2])

is equivalent to

Compose [Scaled m1 (zonk t1), Scaled m2 (zonk t2)]

while we need Scaled (zonk m1) (zonk t1).

It turned out that every place using Compose was
incorrectly ignoring multiplicities.

Instead of substTys that is polymorphic in Traversable t,
we're back to monomorphic substTys that takes only a list
and a new function substScaledTys. I removed unneccessary
polymorphism to make the patch smaller, and the Traversable
instance so that it won't tempt us again.

The test MultConstructor is not fully working - the function
g causes a Lint error. This is a different bug though.