Make all submodules have absolute URLs
The relative URLs were a workaround to let most contributors fork from
Github due to a weakness in the haskell.org server.

This workaround is no longer needed. And relative submodule URLs are
an impediment to forking which makes contributions harder than they
should be.

The URLs are chosen to clone from https, because this makes sure that
anybody, even not a registered Gitlab user, can clone a fork
recursively.

Conflicts:
	.gitmodules