Fix case-of-known-constructor linearity
Suppose we have a worker for a strict datatype:

data Bin = Bin !Int

$WBin['Omega]
  = \ (dt_aZk['One]) ->
      case dt_aZk['One] of (dt_aZl['One])
      { GHC.Types.I# (dt_aZm['One]) -> Bin['Omega] dt_aZm['One]
      }

Then, $WBin['Omega] k_aYO['Omega] will reply "yes"
to exprIsConApp_maybe: given a case float
CASE k_aYO['Omega] of dt_aZl['One]
   I# [dt_aZm['One]]
this ia an application of constructor Bin to argument dt_aZm.

Then, without the scaling, we would transform

      case TUnboxer.$WBin['Omega] k_aYO['Omega]
      of (wild_00['Omega] ::['Omega] IntPSQ)
      { __DEFAULT -> () }

to

      case k_aYO['Omega]
      of (dt_aZl['One])
      { I# (dt_aZm['One]) -> () }

This is wrong: the argument is not used linearly.
We need to scale the float by multiplicity of the case binder.